+++ example/lib/app/app_module.dart
@@ -1,7 +1,7 @@
+import 'package:example/app/app_bloc.dart';
 import 'package:bloc_pattern/bloc_pattern.dart';
 import 'package:flutter/material.dart';
 import 'package:example/app/app_widget.dart';
-import 'package:example/app/app_bloc.dart';
 
 class AppModule extends ModuleWidget {
   @override