+++ lib/src/command/generate_command.dart
@@ -1,5 +1,6 @@
 import 'package:slidy/slidy.dart';
 import 'package:slidy/src/command/sub_command/generate_bloc.dart';
+import 'package:slidy/src/command/sub_command/generate_test_sub_command.dart';
 
 class GenerateCommand extends CommandBase {
   final name = "generate";
@@ -18,6 +19,8 @@ class GenerateCommand extends CommandBase {
     addSubcommand(GenerateBlocAbbrSubCommand());
     addSubcommand(GenerateRepositorySubCommand());
     addSubcommand(GenerateRepositoryAbbrSubCommand());
+    addSubcommand(GenerateTestSubCommand());
+    addSubcommand(GenerateTestAbbrSubCommand());
   }
 }