+++ lib/src/modules/generate.dart
@@ -1,6 +1,11 @@
+import 'dart:io';
+
 import 'package:path/path.dart';
 import 'package:slidy/src/templates/templates.dart' as templates;
 import 'package:slidy/src/utils/file_utils.dart' as file_utils;
+import 'package:slidy/src/utils/file_utils.dart';
+import 'package:slidy/src/utils/utils.dart';
+import 'package:slidy/src/utils/output_utils.dart' as output;
 
 class Generate {
   static module(String path, bool createCompleteModule) async {
@@ -33,11 +38,70 @@ class Generate {
     }
   }
 
-  static repository(String path) {
-    file_utils.createFile(path, 'repository', templates.repositoryGenerator);
+  static test(String path) {
+    if (path.contains(".dart")) {
+      File entity = File(libPath(path));
+      if (!entity.existsSync()) {
+        output.error("File $path not exist");
+        exit(1);
+      }
+      _generateTest(
+          entity,
+          File(libPath(path)
+              .replaceFirst("lib/", "test/")
+              .replaceFirst(".dart", "_test.dart")));
+    } else {
+      Directory entity = Directory(libPath(path));
+      if (!entity.existsSync()) {
+        output.error("Directory $path not exist");
+        exit(1);
+      }
+
+      for (FileSystemEntity file in entity.listSync()) {
+        if (file is File) {
+          _generateTest(
+              file,
+              File(file.path
+                  .replaceFirst("lib/", "test/")
+                  .replaceFirst(".dart", "_test.dart")));
+        }
+      }
+    }
   }
 
-  static bloc(String path) {
-    file_utils.createFile(path, 'bloc', templates.blocGenerator);
+  static _generateTest(File entity, File entityTest) async {
+    if (entityTest.existsSync()) {
+      output.error("Test already exists");
+      exit(1);
+    }
+
+    String name = basename(entity.path);
+
+    if (name.contains("_bloc.dart")) {
+      entityTest.createSync(recursive: true);
+      output.msg("File test ${entityTest.path} created");
+      entityTest.writeAsStringSync(templates.blocTestGenerator(
+          formatName(name.replaceFirst("_bloc.dart", "")),
+          await getNamePackage(),
+          entity.path));
+    } else if (name.contains("_repository.dart")) {
+      entityTest.createSync(recursive: true);
+      output.msg("File test ${entityTest.path} created");
+      entityTest.writeAsStringSync(templates.repositoryTestGenerator(
+          formatName(name.replaceFirst("_repository.dart", "")),
+          await getNamePackage(),
+          entity.path));
+    }
+    formatFile(entityTest);
+  }
+
+  static repository(String path, [bool isTest = true]) {
+    file_utils.createFile(path, 'repository', templates.repositoryGenerator,
+        isTest ? templates.repositoryTestGenerator : null);
+  }
+
+  static bloc(String path, [bool isTest = true]) {
+    file_utils.createFile(path, 'bloc', templates.blocGenerator,
+        isTest ? templates.blocTestGenerator : null);
   }
 }