+++ lib/src/templates/generator/repository_generator.dart
@@ -1,9 +1,18 @@
 String repositoryGenerator(String name) => '''
 import 'package:bloc_pattern/bloc_pattern.dart';
+import 'package:dio/dio.dart';
 
 class ${name}Repository extends Disposable {
 
-  //dispose will be called automatically by closing its streams
+
+  Future fetchPost(Dio client) async {
+    final response =
+        await client.get('https://jsonplaceholder.typicode.com/posts/1');
+    return response.data;
+  }
+
+
+  //dispose will be called automatically
   @override
   void dispose() {