+++ lib/src/templates/start/src/app_module.dart
@@ -2,13 +2,10 @@ String startAppModule(String pkg) => '''
 import 'package:bloc_pattern/bloc_pattern.dart';
 import 'package:flutter/material.dart';
 import 'package:${pkg}/app/app_widget.dart';
-import 'package:${pkg}/app/app_bloc.dart';
 
 class AppModule extends ModuleWidget {
   @override
-  List<Bloc> get blocs => [
-    Bloc((i) => AppBloc()),
-  ];
+  List<Bloc> get blocs => [];
 
   @override
   List<Dependency> get dependencies => [];